fix(docker): Double-quote `REDIS_VERSION` Docker ARG
This is related to https://github.com/fly-apps/redis/pull/18, in the
context of resolving double-quote escape issues.

There's no known report of this Dockerfile being an issue in the wild,
but given https://github.com/fly-apps/redis/pull/18, I felt it would be
a good idea to nip it in the bud.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>